Revert "Splited test from HyperdisksIT"
This reverts commit d3a06cc269e32a10a0fed6dd8a938c8e24e332c8.
