+++ compute/cloud-client/src/test/java/compute/disks/HyperdisksIT.java
@@ -19,6 +19,10 @@ package compute.disks;
 import static com.google.common.truth.Truth.assertWithMessage;
 
 import com.google.cloud.compute.v1.Disk;
+import com.google.cloud.compute.v1.StoragePool;
+import com.google.cloud.compute.v1.StoragePoolsClient;
+import compute.disks.storagepool.CreateDiskInStoragePool;
+import compute.disks.storagepool.CreateHyperdiskStoragePool;
 import java.io.IOException;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
@@ -39,8 +43,13 @@ import org.junit.runners.MethodSorters;
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class HyperdisksIT {
   private static final String PROJECT_ID = System.getenv("GOOGLE_CLOUD_PROJECT");
-  private static final String ZONE = "europe-west1-b";
+  // Zone in which the hyperdisk will be created.
+  private static final String ZONE_1 = "europe-west1-b";
+  // Zone in which the storage pool will be created.
+  private static final String ZONE_2 = "us-central1-a";
   private static String HYPERDISK_NAME;
+  private static String HYPERDISK_IN_POOL_NAME;
+  private static String STORAGE_POOL_NAME;
 
   // Check if the required environment variables are set.
   public static void requireEnvVar(String envVarName) {
@@ -54,22 +63,29 @@ public class HyperdisksIT {
     requireEnvVar("GOOGLE_CLOUD_PROJECT");
 
     HYPERDISK_NAME = "test-hyperdisk-enc-" + UUID.randomUUID();
+    HYPERDISK_IN_POOL_NAME = "test-hyperdisk-enc-" + UUID.randomUUID();
+    STORAGE_POOL_NAME = "test-storage-pool-enc-" + UUID.randomUUID();
   }
 
   @AfterAll
   public static void cleanup()
        throws IOException, InterruptedException, ExecutionException, TimeoutException {
     // Delete all disks created for testing.
-    DeleteDisk.deleteDisk(PROJECT_ID, ZONE, HYPERDISK_NAME);
+    DeleteDisk.deleteDisk(PROJECT_ID, ZONE_1, HYPERDISK_NAME);
+    DeleteDisk.deleteDisk(PROJECT_ID, ZONE_2, HYPERDISK_IN_POOL_NAME);
+
+    try (StoragePoolsClient client = StoragePoolsClient.create()) {
+      client.deleteAsync(PROJECT_ID, ZONE_2, STORAGE_POOL_NAME);
+    }
   }
 
   @Test
   public void stage1_CreateHyperdiskTest()
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    String diskType = String.format("zones/%s/diskTypes/hyperdisk-balanced", ZONE);
+    String diskType = String.format("zones/%s/diskTypes/hyperdisk-balanced", ZONE_1);
 
     Disk hyperdisk = CreateHyperdisk
-        .createHyperdisk(PROJECT_ID, ZONE, HYPERDISK_NAME, diskType,
+        .createHyperdisk(PROJECT_ID, ZONE_1, HYPERDISK_NAME, diskType,
             10, 3000, 140);
 
     Assert.assertNotNull(hyperdisk);
@@ -78,6 +94,53 @@ public class HyperdisksIT {
     Assert.assertEquals(140, hyperdisk.getProvisionedThroughput());
     Assert.assertEquals(10, hyperdisk.getSizeGb());
     Assert.assertTrue(hyperdisk.getType().contains("hyperdisk-balanced"));
-    Assert.assertTrue(hyperdisk.getZone().contains(ZONE));
+    Assert.assertTrue(hyperdisk.getZone().contains(ZONE_1));
+  }
+
+  @Test
+  public void stage1_CreateHyperdiskStoragePoolTest()
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+    String poolType = String.format("projects/%s/zones/%s/storagePoolTypes/hyperdisk-balanced",
+        PROJECT_ID, ZONE_2);
+    StoragePool storagePool = CreateHyperdiskStoragePool
+         .createHyperdiskStoragePool(PROJECT_ID, ZONE_2, STORAGE_POOL_NAME, poolType,
+         "advanced", 10240, 10000, 10240);
+
+    // Wait for server update
+    TimeUnit.MINUTES.sleep(3);
+
+    Assert.assertNotNull(storagePool);
+    Assert.assertEquals(STORAGE_POOL_NAME, storagePool.getName());
+    Assert.assertEquals(10000, storagePool.getPoolProvisionedIops());
+    Assert.assertEquals(10240, storagePool.getPoolProvisionedThroughput());
+    Assert.assertEquals(10240, storagePool.getPoolProvisionedCapacityGb());
+    Assert.assertTrue(storagePool.getStoragePoolType().contains("hyperdisk-balanced"));
+    Assert.assertTrue(storagePool.getCapacityProvisioningType().equalsIgnoreCase("advanced"));
+    Assert.assertTrue(storagePool.getZone().contains(ZONE_2));
+  }
+
+  @Test
+  public void stage2_CreateHyperdiskStoragePoolTest()
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+    String diskType = String.format("zones/%s/diskTypes/hyperdisk-balanced", ZONE_2);
+    String storagePoolLink = String
+        .format("https://www.googleapis.com/compute/v1/projects/%s/zones/%s/storagePools/%s",
+           PROJECT_ID, ZONE_2, STORAGE_POOL_NAME);
+
+    Disk disk = CreateDiskInStoragePool
+        .createDiskInStoragePool(PROJECT_ID, ZONE_2, HYPERDISK_IN_POOL_NAME, storagePoolLink,
+           diskType, 10, 3000, 140);
+
+    // Wait for server update
+    TimeUnit.MINUTES.sleep(3);
+
+    Assert.assertNotNull(disk);
+    Assert.assertEquals(HYPERDISK_IN_POOL_NAME, disk.getName());
+    Assert.assertTrue(disk.getStoragePool().contains(STORAGE_POOL_NAME));
+    Assert.assertEquals(3000, disk.getProvisionedIops());
+    Assert.assertEquals(140, disk.getProvisionedThroughput());
+    Assert.assertEquals(10, disk.getSizeGb());
+    Assert.assertTrue(disk.getType().contains("hyperdisk-balanced"));
+    Assert.assertTrue(disk.getZone().contains(ZONE_2));
   }
 }
